requirements.txt
ipykernel
numpy
pandas
matplotlib
google-generativeai
python-dotenv
scipy